+++ ui/profiles/profileDetailsStore.tsx
@@ -1,3 +1,4 @@
+import { ExpandedProfile } from '@/features/pocketbase/stores/types'
 import { createContext, ReactNode, useState } from 'react'
 import { createStore, StoreApi } from 'zustand'
 
@@ -10,6 +11,7 @@ type ProfileDetailsState = {
   setIsEditing: (newValue: boolean) => void
   openedFromFeed: boolean
   editingRights: boolean
+  profile: ExpandedProfile
 }
 
 // define a "profile store"
@@ -23,11 +25,13 @@ export const ProfileDetailsProvider = ({
   initialIndex,
   openedFromFeed,
   editingRights,
+  profile,
 }: {
   children: ReactNode
   initialIndex: number
   openedFromFeed: boolean
   editingRights: boolean
+  profile: ExpandedProfile
 }) => {
   const [profileDetailsStore] = useState(() =>
     createStore<ProfileDetailsState>((set) => ({
@@ -37,6 +41,7 @@ export const ProfileDetailsProvider = ({
       setShowContextMenu: (newShowContextMenu) => set({ showContextMenu: newShowContextMenu }),
       isEditing: false,
       setIsEditing: (newIsEditing) => set({ isEditing: newIsEditing }),
+      profile,
       openedFromFeed,
       editingRights,
     }))
